// user-dashboard/store/user/action-type.ts
export enum ActionTypes {
  GET_USERS = "GET_USERS",
  GET_USER = "GET_USER",
  CHECK_USER_SLUG_AVAILABLE = "CHECK_USER_SLUG_AVAILABLE",
  CHECK_USERNAME_AVAILABLE = "CHECK_USERNAME_AVAILABLE",
  GET_USER_BY_SLUG = "GET_USER_BY_SLUG",
  UPDATE_USER = "UPDATE_USER",
  PUBLISH_USER = "PUBLISH_USER",
  HARD_DELETE_USER = "HARD_DELETE_USER",
  UPDATE_USER_AVATAR = "UPDATE_USER_AVATAR",
}